--- {{cookiecutter.repo_name}}/notes.txt
@@ -1,36 +1,0 @@
-documentation notes:
-Security
-
-- http only cookie
-
-
-
-mention that you need to use python3.8 to manage.py on prod
-
-
-Truly Static assets - where do they live
- - Django -  easier to put a cdn in front of them.
-Get cdn configured with next and django
-https://nextjs.org/docs/api-reference/next.config.js/cdn-support-with-asset-prefix
-
-fix security issues from:
-https://securityheaders.com/?q=https%3A%2F%2Flightmatter-{{cookiecutter.repo_name}}.herokuapp.com%2F&followRedirects=on
-
-Better jest testing:
-https://github.com/testing-library/jest-dom
-
-
-cypress testing
-https://github.com/cypress-io/cypress-example-recipes
-- https://github.com/bahmutov/cypress-select-tests
-- get client side mock working w/ mirage, but not enabled when running tests through django
-- move nextjs server start into a once per class thing vs once per test in the django test case
-- figure out how to enforce new build of nextjs on running django tests
-- https://github.com/svish/cypress-hmr-restarter#readme for interactive testing
-   - https://stackoverflow.com/questions/52231111/re-run-cypress-tests-in-gui-when-webpack-dev-server-causes-page-reload
-
-
-Sentry - Get consolidated error report for frontend and backend
- - Figure out what to do about release number
-https://nextjs.org/docs/advanced-features/measuring-performance#sending-results-to-analytics
-https://github.com/zeit/next.js/discussions/12913 -- get error page working
